chore: split the lib into multiple files to segment responsibilities and improve code organization
