+++ response_cache_writer.go
@@ -1,0 +1,24 @@
+package cache
+
+import (
+	"bytes"
+
+	"github.com/gin-gonic/gin"
+)
+
+// responseCacheWriter
+type responseCacheWriter struct {
+	gin.ResponseWriter
+
+	body bytes.Buffer
+}
+
+func (w *responseCacheWriter) Write(b []byte) (int, error) {
+	w.body.Write(b)
+	return w.ResponseWriter.Write(b)
+}
+
+func (w *responseCacheWriter) WriteString(s string) (int, error) {
+	w.body.WriteString(s)
+	return w.ResponseWriter.WriteString(s)
+}
